refactor(ui): shrink the sidebar's scroll-edge hook to sticky insets (#333)
The shadcn `scroll-fade` utility drives its ramp depth off a scroll timeline,
so nothing needs the has-content-above/below booleans any more. What survives
is the sticky-header measurement: the ramps still have to start below whatever
section headers are parked at the edge, and that stack changes as you scroll,
so the listener stays.

The measurement itself is now a pure `stickyInset(bounds, headers, edge)` the
unit test can exercise without layout.
